# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.ChangePassword:
    properties:
      email:
        type: string
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
...
  models.EducationReturnValue:
    properties:
      endYear:
        type: integer
      id:
        type: integer
      name:
        type: string
      startYear:
        type: integer
    type: object
...
  models.ExperienceReturnValue:
    properties:
      endYear:
        type: integer
      id:
        type: integer
      place:
        type: string
      position:
        type: string
      startYear:
        type: integer
    type: object
...
  models.LoginUserPassword:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.PortfolioDatabase:
    properties:
      description:
        type: string
      id:
        type: string
      link:
        type: string
      ownerId:
        type: string
    type: object
...
  models.RegistrationUserUsingPassword:
    properties:
      description:
        type: string
      email:
        type: string
      fullname:
        type: string
      location:
        type: integer
      password:
        type: string
      skills:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  models.ResetPassword:
    properties:
      email:
        type: string
    type: object
...
  models.ResponseOKGetUserProfile:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/models.UserProfile'
        type: object
      message:
        type: string
    type: object
...
  models.ResponseOKLogin:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/models.TokenResponse'
        type: object
      message:
        type: string
    type: object
...
  models.ResponseWithNoBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  models.TokenResponse:
    properties:
      expire:
        type: string
      token:
        type: string
    type: object
  models.UpdateResetPassword:
    properties:
      password:
        type: string
      token:
        type: string
    type: object
...
  models.UserProfile:
    properties:
      description:
        type: string
      education:
        items:
          $ref: '#/definitions/models.EducationReturnValue'
        type: array
      email:
        type: string
      experience:
        items:
          $ref: '#/definitions/models.ExperienceReturnValue'
        type: array
      fullname:
        type: string
      id:
        type: string
      location:
        type: string
      member:
        type: string
      picture:
        type: string
      portfolio:
        items:
          $ref: '#/definitions/models.PortfolioDatabase'
        type: array
      skill:
        items:
          $ref: '#/definitions/models.UserSkills'
        type: array
      username:
        type: string
    type: object
  models.UserSkills:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact:
    name: API Support
  description: Swagger API for Golang Project PMS + Freelancer.
  license: {}
  termsOfService: http://swagger.io/terms/
  title: Swagger API
  version: "1.0"
paths:
...
  /change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Information needed to change password
        in: body
        name: Info
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Register new user using email and password
      tags:
      - User
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account
        in: body
        name: parameter
        required: true
        schema:
          $ref: '#/definitions/models.LoginUserPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOKLogin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Login user using email and password
      tags:
      - User
  /logout:
    get:
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Logout
      tags:
      - User
...
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.RegistrationUserUsingPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Register new user using email and password
      tags:
      - User
  /resetPassword:
    post:
      consumes:
      - application/json
      parameters:
      - description: Information needed to change password
        in: body
        name: Info
        required: true
        schema:
          $ref: '#/definitions/models.ResetPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Reset Password (Forget Password Feature)
      tags:
      - User
...
  /updateNewPassword:
    post:
      consumes:
      - application/json
      parameters:
      - description: Information needed to update password
        in: body
        name: Info
        required: true
        schema:
          $ref: '#/definitions/models.UpdateResetPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: Update Password (after reset password)
      tags:
      - User
...
  /userProfile:
    get:
      consumes:
      - application/json
      parameters:
      - description: Token Header
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOKGetUserProfile'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseWithNoBody'
      summary: User Profile Data
      tags:
      - User
...
swagger: "2.0"